The decimal save pointable was using the wrong type.
git-svn-id: https://svn.apache.org/repos/asf/incubator/vxquery/trunk/vxquery@1369962 13f79535-47bb-0310-9956-ffa450edef68

+++ vxquery-core/src/main/java/org/apache/vxquery/runtime/functions/arithmetic/AddOperation.java
@@ -109,10 +109,11 @@ public class AddOperation extends AbstractArithmeticOperation {
             throw new SystemException(ErrorCode.XPDY0002);
         }
         value1 += value2;
+
         // Save
         dOut.write(ValueTag.XS_DECIMAL_TAG);
         dOut.writeByte(place1);
-        dOut.writeDouble(value1);
+        dOut.writeLong(value1);
     }
 
     @Override
